Major changes in handling users without username.
Now a user is warned three times about the absence of his username.
At the fourth message without username the user permissions are restricted
and he can't write any other message in the chat.
At the same time an InlineButton is sent to him (it is specific for him) so that:
    * the user can set a username
    * the user press the button to check if is has set it correctly
In the case that all is ok, the user standard permission are restored.

Changes to are in these files:
	modified:   conversation/check_text.py
	renamed:    conversation/answer_check.py -> conversation/query_reaction.py
	modified:   conversation/welcome.py
	modified:   the_kirchhoff_bot.py